+++ CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed: Main Language option value for WordPress Setting
 * Fixed: make Date on Tab clickable on Tabbed Schedule View
 * Fixed: prevent possible conflicts with changes not saved reload message
-* Fixed: do not conflict check Shift against self for last shift check
+* Fixed: do not conflict check Shift against itself for last shift check
+* Fixed: link back to Show posts for related Show posts (allow multiple)
+* Fixed: filter next/previous post link for (multiple) related Show posts
 
 ### 2.3.3.5
 * Fixed: use schedule based on start_day if specified for Schedule view